Refactored Charset location
